Add a "enrollment" analysis window
This corresponds to the first daily window, but it allows us to capture some CFR events that should happen on the day of enrollment.

It's possible we should actually wait "a few" days for this, since some impressions could happen at 23:59 and not get a response until the next day, etc.