Run overall metrics
Fix an early return that kept us from running overall metrics, and run them a little earlier.

The "end date" is kind of funny because it represents the date that operators intend to turn off the experiment. The prior day is the last full day of data, so we can actually compute our final results before the experiment gets switched off.