Template enrollment queries (#117)
Allow the enrollment queries to reference properties of the experiment. 
Simplifies the fallback getter for ExperimentConfiguration.

This was probably an exercise in making a simple thing hard, but:
- creates a proxy object to prevent a template from referencing itself 
recursively
- caches the result of the template expansion at instantiation to make 
sure that any failures happen early

The cached template will be invalid if the ExperimentSpec or 
Experimenter object are mutated after the config spec is resolved, so, 
uh, hopefully nobody will try that.