Add a count statistic (#121)
Add a special statistic that counts the number of clients enrolled in each branch. It doesn't depend on the value of any of the metrics columns -- just the number of rows for each branch, since each metrics table will have one row per client per branch.

It's awkward to express in configuration since it doesn't have a corresponding metric and it seems in poor form to invent one.

It should be the same for all statistics tables (but may not be, because of Shredder).