Fix KDEs (#312)
What happened here is that I had conflated the number of points to use while fitting the KDE with the number of points in the result. Using a larger number of points to fit and then downsampling to an arbitrary scale gives a better representation of the function, both because the fit has higher fidelity and the result uses the output space better.

* Break import cycles

* Factor out make_grid

* Use better grids for KDEs

Construct a kernel estimate using a larger grid and then sample it down as needed. Also supports sampling onto a geometric grid.

* Add a zero point to the grid if necessary

* Add tests for make_grid