Fix Fenix experiment analysis (#578)
I misunderstood how cattrs and attrs interacted with respect to _prefixed attributes.

attrs expects the attributes to appear in the constructor without a prefix. cattrs expects the attributes to appear in the dictionary with a prefix!

```python
>>> import cattr
>>> import attr
>>> @attr.s(auto_attribs=True)
... class A:
...   _foo: str
...
>>> A(foo="bar")
A(_foo='bar')
>>> cattr.structure({"_foo": "bar"}, A)
A(_foo='bar')
>>> cattr.structure({"foo": "bar"}, A)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/tsmith/projects/jetstream/venv/lib/python3.9/site-packages/cattr/converters.py", line 205, in structure
    return self._structure_func.dispatch(cl)(obj, cl)
  File "/Users/tsmith/projects/jetstream/venv/lib/python3.9/site-packages/cattr/converters.py", line 345, in structure_attrs_fromdict
    return cl(**conv_obj)  # type: ignore
TypeError: __init__() missing 1 required positional argument: 'foo'
```

This was causing us to fallback to firefox-desktop in the property accessors, which meant that the experiment was treated by mozanalysis as a desktop experiment, which meant we didn't observe any enrollments!

This fix is suggested at https://github.com/ Tinche/cattrs/issues/23#issuecomment-776655547.