Use metrics table for total URI count
mobile_search_clients_engines_sources_daily is not a good source for this because total URI count is repeated for each (engine, source) search pair.

Using the metrics table without a filter for foreground activity should be correct for this because the URI count should always be zero in the absence of foreground activity.